Migrate to SpellEngine 1.10 and replace Accessories with Curios
This mod sat on SpellEngine 1.8.10, the oldest base in the ecosystem,
but has no particle code, no SpellHelper usage and no DescriptionMutator
- so the version bump alone compiles clean and the slot-mod swap is the
whole migration.

Dependencies: spell_engine 1.8.10 -> 1.10.1.001, more_rpg_library 2.5.0
-> 2.6.5.003, both from mavenLocal with transitive = false; spell_power
1.4.3 -> 1.6.0.

Accessories is gone from SpellEngine 1.10; the ecosystem now runs
Trinkets on Fabric and Curios 9.5.1 native on NeoForge. Common ends up
with zero slot-mod references: AdditionalJewelryFactory.factory is the
seam, overwritten by TrinketsHelper on Fabric and the new CuriosHelper
on NeoForge, each from a module only that platform compiles. Slot
membership is untouched and still data-driven - this mod's items enter
#jewelry:rings / #jewelry:necklaces, and Jewelry's own per-platform tags
point at those.

The jewelry_version bump 2.3.1 -> 2.4.0 is load-bearing, not cosmetic.
This is an add-on built out of Jewelry's item classes: 2.3.1 ships
JewelryAccessoriesItem and no Curios anything, while 2.4.0 ships
JewelryCurioItem. NeoForge now reuses that class exactly as Fabric
already reuses JewelryTrinketItem, which inherits Jewelry's
getAttributeModifiers override (Curios ignores AttributeModifiersComponent
on Item.Settings), its per-slot-and-per-item modifier id guarding against
"Modifier is already applied" on fast swaps, and its equip-sound
handling - rather than duplicating all three here.

Gated on net.jewelry.Platform, not SpellEngine's: after this change the
mod has zero spell_engine references and neither manifest declares SE, so
SE is optional at runtime and class-loading its Platform would be unsafe.
jewelry is a required dependency, so its Platform is always present.

API drift the compiler could not see until the bump: Jewelry 2.4.0
deleted net.jewelry.fabric.compat.SpellEngineHelper, whose
initSlotCompat() the old Fabric CompatFeatures called to ask SE which
slot mod to prefer - a negotiation that stopped existing once there was
one slot mod per loader. That call sat inside a catch (Exception), which
would NOT have caught the resulting NoClassDefFoundError, since that is
an Error. On a 2.4.0 runtime the old code crashed mod init outright.

Also dropped FabricCompatConfig, whose only field chose between Trinkets
and Accessories, mirroring Jewelry 2.4.0 deleting its equivalent.

Kept deliberately, against the precedent set in the Witcher migration:
- forgeRuntimeLibrary("blue.endless:jankson") - Jewelry 2.4.0, the closer
  analogue, keeps that line beside its Curios dep. Dev-classpath only.
- the su5ed maven, recommented - it was filed under an Accessories
  comment but is the only host for forgified-fabric-api, which the
  neoforge module still needs. Removing it broke resolution.

Made gradlew executable; it shipped mode 644, which silently produced a
false-green compile probe.

:neoforge:dependencies confirms curios present with zero
wispforest/accessories entries. Datagen: zero deletions. Both platforms
compile. Not runtime-tested - in particular that Curios applies the
modifiers in the ring/necklace slots, which is exactly what the
AttributeModifiersComponent workaround exists for.

Two things left undecided, deliberately:
- fabric.mod.json still floors jewelry at >=1.3.0. Nothing on Fabric
  forces 2.4.0, but a user on 2.3.x could see Jewelry's own items routed
  to Accessories while this mod's go to Trinkets.
- neoforge.mods.toml orders this mod BEFORE jewelry, which reads
  backwards for an add-on. Pre-existing and harmless to the Curios path,
  so left alone.
