--- common/src/main/java/net/additional_jewelry/compat/AccessoriesCompat.java
@@ -1,12 +1,0 @@
-package net.additional_jewelry.compat;
-
-import net.fabricmc.loader.api.FabricLoader;
-
-public class AccessoriesCompat {
-    public static void init() {
-        if (FabricLoader.getInstance().isModLoaded("accessories")) {
-            // Outsource to avoid class loading issues
-            AccessoriesHelper.registerFactory();
-        }
-    }
-}
