--- fabric/src/main/java/com/additional_jewelry/fabric/compat/FabricCompatConfig.java
@@ -1,6 +1,0 @@
-package com.additional_jewelry.fabric.compat;
-
-public class FabricCompatConfig {
-    public String disclaimer = "THIS CONFIG DOES NOT SYNC FROM SERVER TO CLIENT. PLEASE MAKE SURE TO HAVE THE SAME CONFIG ON BOTH SIDES.";
-    public String preferred_slot_mod = "trinkets";
-}
