+++ gradle.properties
@@ -21,20 +21,18 @@ forgified_fabric_api_version=0.115.6+2.1.4+1.21.1
 fabric_version=0.115.6+1.21.1
 cloth_config_version=15.0.130
 player_anim_version=2.0.1+1.21.1
-spell_power_version=1.4.3+1.21.1
+spell_power_version=1.6.0+1.21.1
 tiny_config_version=3.1.0
 mixin_extras_version=0.1.1
-spell_engine_version=1.8.10+1.21.1
+spell_engine_version=1.10.1.001+1.21.1
 structure_pool_api_version=1.2.0+1.21.1
-more_rpg_library_version=2.5.0+1.21.1
+more_rpg_library_version=2.6.5.003+1.21.1
 ranged_weapon_api_version=2.3.2+1.21.1
-jewelry_version =2.3.1+1.21.1
+jewelry_version =2.4.0+1.21.1
 
 # Fabric only
 trinkets_version=3.10.0
 mod_menu_version=11.0.2
 
 # NeoForge only
-owo_version=0.12.15.4+1.21
-owo_neo_version=0.12.15.5-beta.1+1.21
-accessories_version=1.1.0-beta.48+1.21.1
+curios_version=9.5.1+1.21.1
