+++ neoforge/build.gradle
@@ -59,8 +59,9 @@ repositories {
         url = 'https://maven.ladysnake.org/releases'
     }
 
-    // For Accessories
-    maven { url 'https://maven.wispforest.io/releases' }
+    // For Curios
+    maven { url 'https://maven.theillusivec4.top/' }
+    // For Forgified Fabric API
     maven { url 'https://maven.su5ed.dev/releases' }
     maven { url 'https://maven.fabricmc.net' }
 
@@ -75,21 +76,16 @@ dependencies {
 
     modImplementation("maven.modrinth:tiny-config:${project.tiny_config_version}-${platform}")
 
-    // Accessories - https://docs.wispforest.io/accessories/developer/dev_setup
-    modImplementation "maven.modrinth:owo-lib:${project.owo_neo_version}" // AccessoryItem wouldn't compile without this
-    modImplementation("io.wispforest:accessories-neoforge:${project.accessories_version}")
+    // Curios - https://docs.illusivesoulworks.com/category/curios
+    modImplementation("top.theillusivec4.curios:curios-neoforge:${project.curios_version}")
     forgeRuntimeLibrary("blue.endless:jankson:1.2.2")
-    forgeRuntimeLibrary("io.wispforest:endec:0.1.5.1")
-    forgeRuntimeLibrary("io.wispforest.endec:gson:0.1.3.1")
-    forgeRuntimeLibrary("io.wispforest.endec:jankson:0.1.3.1")
-    forgeRuntimeLibrary("io.wispforest.endec:netty:0.1.2")
 
     modImplementation("maven.modrinth:structure-pool-api:${project.structure_pool_api_version}-${platform}")
     modImplementation("maven.modrinth:spell-power:${project.spell_power_version}-${platform}")
     modImplementation("maven.modrinth:ranged-weapon-api:${project.ranged_weapon_api_version}-${platform}")
     modImplementation("maven.modrinth:jewelry:${project.jewelry_version}-${platform}")
-    modImplementation("maven.modrinth:more-rpg-library:${project.more_rpg_library_version}-${platform}")
-    modImplementation("maven.modrinth:spell-engine:${project.spell_engine_version}-${platform}")
+    modImplementation("net.more_rpg_classes:more_rpg_library-${platform}:${project.more_rpg_library_version}") { transitive = false }
+    modImplementation("net.spell_engine:spell_engine-${platform}:${project.spell_engine_version}") { transitive = false }
     modImplementation("dev.kosmx.player-anim:player-animation-lib-forge:${project.player_anim_version}")
     modImplementation("me.shedaniel.cloth:cloth-config-${platform}:${project.cloth_config_version}")
 }
