+++ neoforge/src/main/java/com/additional_jewelry/neoforge/compat/curios/CuriosCompat.java
@@ -1,0 +1,12 @@
+package com.additional_jewelry.neoforge.compat.curios;
+
+import net.jewelry.Platform;
+
+public class CuriosCompat {
+    public static void init() {
+        if (Platform.util().isModLoaded("curios")) {
+            // Outsource to avoid class loading issues
+            CuriosHelper.registerFactory();
+        }
+    }
+}
