+++ neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -34,6 +34,12 @@ versionRange = "[${jewelry_version},)"
 ordering = "BEFORE"
 side = "BOTH"
 
+[[dependencies.additional_rpg_jewelry]]
+modId = "curios"
+type = "optional"
+ordering = "AFTER"
+side = "BOTH"
+
 
 [[mixins]]
 config = "additional_rpg_jewelry.mixins.json"
